Remove dynamic iskeyword
Set iskeyword locally to buffer in filetype.  This allows us to remove
the over-complicated and fragile dynamic rules for adding or removing
"-" from keyword depending on context.

This removes the original author's desire to be able to use `w`
navigation to parts of words.  To reinstate that, simply install
the plugin at https://github.com/bkad/CamelCaseMotion/ and use
`<Leader>w` etc. instead.